don't recreate backup if one exists

+++ change-exe-icon.sh
@@ -29,7 +29,10 @@ if [[ -f "$DATA_DIR"/game.ico ]]; then
         fi
     fi
 
-    cp "$EXECUTABLE" "$EXECUTABLE.bak"
+    if [[ ! -e "$EXECUTABLE.bak" ]]; then
+        cp "$EXECUTABLE" "$EXECUTABLE.bak"
+    fi
+
     cp "$DATA_DIR/game.ico" "$DATA_DIR/1.ico"
 
     wine resource_hacker/ResourceHacker.exe -open "$EXECUTABLE" -save "$EXECUTABLE" -log ./resourcehacker.log -resource "$DATA_DIR/1.ico" -action addoverwrite -mask ICONGROUP,MAINICON,
